Move `fft` to `math` folder.
