+++ gpu/src/metal/fft/ops.rs
@@ -133,7 +133,7 @@ fn void_ptr<T>(v: &T) -> *const core::ffi::c_void {
 #[cfg(test)]
 mod tests {
     use crate::metal::abstractions::state::*;
-    use lambdaworks_fft::roots_of_unity::get_twiddles;
+    use lambdaworks_math::fft::roots_of_unity::get_twiddles;
     use lambdaworks_math::field::{
         fields::fft_friendly::stark_252_prime_field::Stark252PrimeField, traits::RootsConfig,
     };
@@ -169,7 +169,7 @@ mod tests {
             let twiddles = get_twiddles(order.into(), RootsConfig::BitReverse).unwrap();
 
             let metal_result = super::fft(&input, &twiddles, &metal_state).unwrap();
-            let sequential_result = lambdaworks_fft::ops::fft(&input, &twiddles).unwrap();
+            let sequential_result = lambdaworks_math::fft::ops::fft(&input, &twiddles).unwrap();
 
             prop_assert_eq!(&metal_result, &sequential_result);
         }
