+++ math/src/fft/roots_of_unity.rs
@@ -1,9 +1,9 @@
-use lambdaworks_math::field::{
+use crate::field::{
     element::FieldElement,
     traits::{IsFFTField, RootsConfig},
 };
 
-use crate::{bit_reversing::in_place_bit_reverse_permute, errors::FFTError};
+use crate::{fft::bit_reversing::in_place_bit_reverse_permute, fft::errors::FFTError};
 
 /// Returns a `Vec` of the powers of a `2^n`th primitive root of unity in some configuration
 /// `config`. For example, in a `Natural` config this would yield: w^0, w^1, w^2...
@@ -69,8 +69,8 @@ pub fn get_twiddles<F: IsFFTField>(
 
 #[cfg(test)]
 mod tests {
-    use crate::{bit_reversing::in_place_bit_reverse_permute, roots_of_unity::get_twiddles};
-    use lambdaworks_math::field::{test_fields::u64_test_field::U64TestField, traits::RootsConfig};
+    use crate::fft::{bit_reversing::in_place_bit_reverse_permute, roots_of_unity::get_twiddles};
+    use crate::field::{test_fields::u64_test_field::U64TestField, traits::RootsConfig};
     use proptest::prelude::*;
 
     type F = U64TestField;
