+++ math/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod cyclic_group;
 pub mod elliptic_curve;
 pub mod errors;
+pub mod fft;
 pub mod field;
 pub mod helpers;
 pub mod msm;
